Extracted rooms accumulating logic into a service
