+++ BookingCalendarApi/Services/IRoomsProvider.cs
@@ -1,0 +1,7 @@
+namespace BookingCalendarApi.Services
+{
+    public interface IRoomsProvider
+    {
+        public Task<IEnumerable<FlattenedRoom>> AccumulateAllRoomsAsync(string from, string to);
+    }
+}
